refactor(preview): finish the install* convention in markdown-config.ts (#1908)
createPreviewMarkdown already delegated 7 concerns to installMath /
installCallouts / installDoiAutolink / installHighlight / installWikiLinks /
installNoteTags / installTransclusions, but kept three more concerns inline:
the heading/paragraph/list-item addressability rules, the six fence
renderers, and the fence dispatcher.

Adds two more install* functions, finishing the pattern:
- installAnchors(md) — heading_open (heading-id-for-anchor stamping),
  paragraph_open (^block-id stamping), and list_item_open (task-checkbox
  rendering). None need PreviewMarkdownDeps, matching every other install*
  function's zero-extra-arg signature.
- installFences(md, deps) — the six fence renderers (output/mermaid/vega/
  vega-lite/youtube/runnable/default) plus the dispatcher. Needs deps (the
  collapse/running fence sets, the run-button gate), so its signature carries
  a second argument like the issue itself specifies.

Moved PreviewMarkdownDeps into its own leaf module (markdown-deps.ts) rather
than leaving it defined in markdown-config.ts: fence-plugin.ts needs the
type, and markdown-config.ts needs installFences from fence-plugin.ts — a
type import back into markdown-config.ts would have made an import cycle
(caught by tests/architecture/no-cycles.test.ts, which doesn't special-case
`import type`). markdown-config.ts re-exports the type so both existing test
files that import it from there are unaffected.

The image rule and the :::query-… block directive stay inline — the issue's
own written signatures (installFences(md, deps) but installAnchors(md), no
deps) rule out folding the image rule into installAnchors, and neither is
named in the success criteria.

createPreviewMarkdown: 402 lines of logic -> composition (7 existing calls +
2 new + the image rule + the query directive). markdown-config.ts: 470 -> 198
lines. Existing tests (markdown-config.test.ts, hydrate.test.ts, plus
sanitize-note-html.test.ts) pass unchanged — 106+ tests untouched.
